Neatening source code, adding missing modules
* Removed copyright message encryption, fixed MCI string
* Sorting 1.3's equates file by address

Problem with using {include:} in this convoluted project is that as
multiple modules are included, symbols cannot be resolved properly
since a symbol definition may occur in a file yet to be included.
I tried working around this by generating multiple *.sym files (one
per *.lbl file) but the individual modules have a lot of dependencies.

That's what a Makefile is for, so I started trying to map out dependencies
that way. It has a ways to go, and I still have some errors, but it's
getting there.

In the meantime, I have tried to use C64List/casm's {include:} directive.

The native C64 Fast Assembler dealt with this (and multiple instances
of the SAME LABEL NAME, argh) gracefully enough to generate a binary.
I have no idea how it accomplished this, unless it added internal
decorators like casm's "local label" function does: @label@1, @label@2,
etc.